# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.LoginRequest:
    properties:
      email:
        example: userone@gmail.com
        type: string
      password:
        example: Userone123+
        type: string
    required:
    - email
    - password
    type: object
  dto.RegisterRequest:
    properties:
      email:
        example: userone@gmail.com
        type: string
      password:
        example: Userone123+
        minLength: 3
        type: string
      username:
        example: userone
        minLength: 3
        type: string
    required:
    - email
    - password
    - username
    type: object
  dto.RegisteredUserResponse:
    properties:
      email:
        example: userone@gmail.com
        type: string
      username:
        example: userone
        type: string
    type: object
  dto.TokenResponse:
    properties:
      access_token:
        type: string
      expires_in:
        type: string
      token_type:
        type: string
    type: object
...
  types.ErrorResponse:
    properties:
      error: {}
      message:
        example: Unauthorized
        type: string
      status:
        example: 401
        type: integer
    type: object
  types.Response:
    properties:
      data: {}
      message:
        type: string
      status:
        type: integer
    type: object
host: localhost:5000
info:
  contact:
    email: ilham15suandi@gmail.com
    name: Ilham Suandi
    url: https://github.com/ilhamSuandi
  description: Growy API
  title: Growy API
  version: 1.0.0
paths:
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: login as user
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully Logged In
          schema:
            allOf:
            - $ref: '#/definitions/types.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.TokenResponse'
              type: object
        "400":
          description: request body is invalid
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      summary: login as user
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register new user
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Successfully registered user
          schema:
            allOf:
            - $ref: '#/definitions/types.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.RegisteredUserResponse'
              type: object
        "400":
          description: request body is invalid
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "409":
          description: user exists
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      summary: Register as user
      tags:
      - Auth
  /auth/renew:
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: Successfully Renewed Access Token
          schema:
            allOf:
            - $ref: '#/definitions/types.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.TokenResponse'
              type: object
        "303":
          description: 'Redirect: User must log in again'
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "401":
          description: 'Unauthorized: Invalid or expired refresh token'
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      summary: Renew Access Token
      tags:
      - Auth
...
securityDefinitions:
  BearerAuth:
    description: 'Example Value: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"